more fixes related to clientside movement 1

+++ game/Player.cpp
@@ -114,27 +114,29 @@ const idEventDef EV_Player_LevelTrigger( "levelTrigger" );
 const idEventDef EV_SpectatorTouch( "spectatorTouch", "et" );
 const idEventDef EV_Player_GetIdealWeapon( "getIdealWeapon", NULL, 's' );
 const idEventDef EV_Player_EnableFallDamage( "<enablefalldamage>", NULL); //Added for coop
+const idEventDef EV_Player_EnableReadClientPhysics("<enablereadclientphysics>", NULL); //Added for coop
 
 CLASS_DECLARATION( idActor, idPlayer )
 	EVENT( EV_Player_GetButtons,			idPlayer::Event_GetButtons )
-	EVENT( EV_Player_GetMove,				idPlayer::Event_GetMove )
-	EVENT( EV_Player_GetViewAngles,			idPlayer::Event_GetViewAngles )
-	EVENT( EV_Player_StopFxFov,				idPlayer::Event_StopFxFov )
-	EVENT( EV_Player_EnableWeapon,			idPlayer::Event_EnableWeapon )
-	EVENT( EV_Player_DisableWeapon,			idPlayer::Event_DisableWeapon )
-	EVENT( EV_Player_GetCurrentWeapon,		idPlayer::Event_GetCurrentWeapon )
-	EVENT( EV_Player_GetPreviousWeapon,		idPlayer::Event_GetPreviousWeapon )
-	EVENT( EV_Player_SelectWeapon,			idPlayer::Event_SelectWeapon )
-	EVENT( EV_Player_GetWeaponEntity,		idPlayer::Event_GetWeaponEntity )
-	EVENT( EV_Player_OpenPDA,				idPlayer::Event_OpenPDA )
-	EVENT( EV_Player_InPDA,					idPlayer::Event_InPDA )
-	EVENT( EV_Player_ExitTeleporter,		idPlayer::Event_ExitTeleporter )
-	EVENT( EV_Player_StopAudioLog,			idPlayer::Event_StopAudioLog )
-	EVENT( EV_Player_HideTip,				idPlayer::Event_HideTip )
-	EVENT( EV_Player_LevelTrigger,			idPlayer::Event_LevelTrigger )
-	EVENT( EV_Gibbed,						idPlayer::Event_Gibbed )
-	EVENT( EV_Player_GetIdealWeapon,		idPlayer::Event_GetIdealWeapon )
-	EVENT( EV_Player_EnableFallDamage,		idPlayer::Event_EnableFallDamage)
+	EVENT( EV_Player_GetMove,					idPlayer::Event_GetMove )
+	EVENT( EV_Player_GetViewAngles,				idPlayer::Event_GetViewAngles )
+	EVENT( EV_Player_StopFxFov,					idPlayer::Event_StopFxFov )
+	EVENT( EV_Player_EnableWeapon,				idPlayer::Event_EnableWeapon )
+	EVENT( EV_Player_DisableWeapon,				idPlayer::Event_DisableWeapon )
+	EVENT( EV_Player_GetCurrentWeapon,			idPlayer::Event_GetCurrentWeapon )
+	EVENT( EV_Player_GetPreviousWeapon,			idPlayer::Event_GetPreviousWeapon )
+	EVENT( EV_Player_SelectWeapon,				idPlayer::Event_SelectWeapon )
+	EVENT( EV_Player_GetWeaponEntity,			idPlayer::Event_GetWeaponEntity )
+	EVENT( EV_Player_OpenPDA,					idPlayer::Event_OpenPDA )
+	EVENT( EV_Player_InPDA,						idPlayer::Event_InPDA )
+	EVENT( EV_Player_ExitTeleporter,			idPlayer::Event_ExitTeleporter )
+	EVENT( EV_Player_StopAudioLog,				idPlayer::Event_StopAudioLog )
+	EVENT( EV_Player_HideTip,					idPlayer::Event_HideTip )
+	EVENT( EV_Player_LevelTrigger,				idPlayer::Event_LevelTrigger )
+	EVENT( EV_Gibbed,							idPlayer::Event_Gibbed )
+	EVENT( EV_Player_GetIdealWeapon,			idPlayer::Event_GetIdealWeapon )
+	EVENT( EV_Player_EnableFallDamage,			idPlayer::Event_EnableFallDamage)
+	EVENT(EV_Player_EnableReadClientPhysics,	idPlayer::Event_EnableReadClientPhysics)
 END_CLASS
 
 const int MAX_RESPAWN_TIME = 10000;
@@ -1681,6 +1683,7 @@ void idPlayer::Init( void ) {
 	nextSendPhysicsInfoTime = gameLocal.clientsideTime + PLAYER_CLIENT_SEND_MOVEMENT*2; //disable clientside movement two seconds
 	nextTimeReadHealth = 0; //added for g_clientsideDamage 1
 	clientSpawnedByServer = false;
+	serverReadPlayerPhysics = false;
 	//coop ends
 
 	if ( hud ) {
@@ -2368,9 +2371,14 @@ idPlayer::PrepareForRestart
 ================
 */
 void idPlayer::PrepareForRestart( void ) {
-	if (gameLocal.mpGame.IsGametypeCoopBased() && gameLocal.isClient) {
-		allowClientsideMovement = false;
-		nextSendPhysicsInfoTime = gameLocal.clientsideTime + 4000; //3segs without clientsidemovement
+	if (gameLocal.mpGame.IsGametypeCoopBased()) {
+		if (gameLocal.isServer) {
+			serverReadPlayerPhysics = false;
+		} else {
+			clientSpawnedByServer = false;
+			allowClientsideMovement = false;
+			nextSendPhysicsInfoTime = gameLocal.clientsideTime + 4000; //3segs without clientsidemovement
+		}
 	}
 	ClearPowerUps();
 	Spectate( true );
@@ -2591,6 +2599,9 @@ void idPlayer::SpawnToPoint( const idVec3 &spawn_origin, const idAngles &spawn_a
 
 
 	if (gameLocal.mpGame.IsGametypeCoopBased()) {
+		CancelEvents(&EV_Player_EnableReadClientPhysics);
+		PostEventSec(&EV_Player_EnableReadClientPhysics, 5.0f);
+		serverReadPlayerPhysics = false;
 		ServerSendEvent( EVENT_PLAYERSPAWN, NULL, false, -1);
 	}
 
@@ -3549,7 +3560,6 @@ void idPlayer::UpdatePowerUps( void ) {
 		assert( !gameLocal.isClient );	// healthPool never be set on client
 		int amt = ( healthPool > 5 ) ? 5 : healthPool;
 		health += amt;
-		common->Printf("Giving to ::UpdatePowerUps\n");
 		if ( health > inventory.maxHealth ) {
 			health = inventory.maxHealth;
 			healthPool = 0;
@@ -4493,7 +4503,7 @@ void idPlayer::UpdateWeapon( void ) {
 			weapon.GetEntity()->GetWeaponDef( animPrefix, inventory.clip[ idealWeapon ] );
 
 			if (gameLocal.isClient && gameLocal.mpGame.IsGametypeCoopBased() && !weapon.GetEntity()->IsLinked()) {
-				gameLocal.Warning("[FATAL]: Avoid crash at idPlayer::UpdateWeapon (2)\n");
+				gameLocal.DWarning("[FATAL]: Avoid crash at idPlayer::UpdateWeapon (2)\n");
 				return; //More duct tape :(
 			}
 
@@ -9111,19 +9121,28 @@ bool idPlayer::ServerReceiveEvent( int event, int time, const idBitMsg &msg ) {
 			return true;
 		}
 		case EVENT_PLAYERPHYSICS: {
-			bool clientsideTeleported;
-			physicsObj.ReadFromEvent(msg);
-			deltaViewAngles[0] = msg.ReadDeltaFloat( 0.0f );
-			deltaViewAngles[1] = msg.ReadDeltaFloat( 0.0f );
-			deltaViewAngles[2] = msg.ReadDeltaFloat( 0.0f );
-			clientsideTeleported = msg.ReadBits(1) != 0;
-			if (clientsideTeleported || clientTeleported) { // to avoid bug related to fall damage kill client with net_clientsideMovement 1
-				noFallDamage = true;
-				CancelEvents(&EV_Player_EnableFallDamage);
-				PostEventSec(&EV_Player_EnableFallDamage, 5.0f);
-				GetPhysics()->SetOrigin(physicsObj.GetOrigin());
-				clientTeleported = false;
-				Move();
+			if (serverReadPlayerPhysics) {
+				bool clientsideTeleported;
+				physicsObj.ReadFromEvent(msg);
+				deltaViewAngles[0] = msg.ReadDeltaFloat(0.0f);
+				deltaViewAngles[1] = msg.ReadDeltaFloat(0.0f);
+				deltaViewAngles[2] = msg.ReadDeltaFloat(0.0f);
+				clientsideTeleported = msg.ReadBits(1) != 0;
+				if (clientsideTeleported || clientTeleported) { // to avoid bug related to fall damage kill client with net_clientsideMovement 1
+					noFallDamage = true;
+					CancelEvents(&EV_Player_EnableFallDamage);
+					PostEventSec(&EV_Player_EnableFallDamage, 5.0f);
+					GetPhysics()->SetOrigin(physicsObj.GetOrigin());
+					clientTeleported = false;
+					Move();
+				}
+			}
+			else {
+				physicsObj.ClearFromEvent(msg);
+				msg.ReadDeltaFloat(0.0f);
+				msg.ReadDeltaFloat(0.0f);
+				msg.ReadDeltaFloat(0.0f);
+				msg.ReadBits(1);
 			}
 			return true;
 		}
@@ -9217,30 +9236,7 @@ bool idPlayer::ClientReceiveEvent( int event, int time, const idBitMsg &msg ) {
 			break;
 		}
 		case EVENT_PLAYERSPAWN: {
-			if (net_clientSideMovement.GetBool()) {
-				if (gameLocal.localClientNum == this->entityNumber) {
-					allowClientsideMovement = true;  //hack for clientsidemovement
-					nextSendPhysicsInfoTime = gameLocal.clientsideTime; //hack for clientsidemovement
-				}
-				idVec3	tmpOrigin = vec3_zero;
-				tmpOrigin.x = msg.ReadFloat();
-				tmpOrigin.y = msg.ReadFloat();
-				tmpOrigin.z = msg.ReadFloat();
-				deltaViewAngles[0] = msg.ReadDeltaFloat( 0.0f );
-				deltaViewAngles[1] = msg.ReadDeltaFloat( 0.0f );
-				deltaViewAngles[2] = msg.ReadDeltaFloat( 0.0f );
-
-				clientTeleported = true;
-				SetOrigin(	tmpOrigin );
-				Move();
-			} else {
-				msg.ReadFloat();
-				msg.ReadFloat();
-				msg.ReadFloat();
-				msg.ReadDeltaFloat( 0.0f );
-				msg.ReadDeltaFloat( 0.0f );
-				msg.ReadDeltaFloat( 0.0f );
-			}
+			clientSpawnedByServer = true;
 			return true;
 			
 		}
@@ -10062,6 +10058,18 @@ idPlayer::Event_EnableFallDamage
 
 void idPlayer::Event_EnableFallDamage(void) {
 	noFallDamage = false;
+}
+
+/*
+================
+idPlayer::Event_EnableReadClientPhysics
+================
+*/
+
+void idPlayer::Event_EnableReadClientPhysics(void) {
+
+	assert(!gameLocal.isClient);
+	serverReadPlayerPhysics = true;
 }
